Support importing URL for tv
Also deprecates discovering from NFC/Android Beam due to deprecation in Android Q.
A follow-up commit will remove this feature completely.
