+++ mobile/src/main/java/com/github/shadowsocks/MainActivity.kt
@@ -23,20 +23,15 @@ package com.github.shadowsocks
 import android.app.Activity
 import android.app.backup.BackupManager
 import android.content.ActivityNotFoundException
-import android.content.DialogInterface
 import android.content.Intent
 import android.net.VpnService
-import android.nfc.NdefMessage
-import android.nfc.NfcAdapter
 import android.os.Bundle
 import android.os.DeadObjectException
 import android.os.Handler
-import android.os.Parcelable
 import android.util.Log
 import android.view.KeyCharacterMap
 import android.view.KeyEvent
 import android.view.MenuItem
-import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.browser.customtabs.CustomTabsIntent
 import androidx.coordinatorlayout.widget.CoordinatorLayout
@@ -51,10 +46,6 @@ import com.github.shadowsocks.aidl.IShadowsocksService
 import com.github.shadowsocks.aidl.ShadowsocksConnection
 import com.github.shadowsocks.aidl.TrafficStats
 import com.github.shadowsocks.bg.BaseService
-import com.github.shadowsocks.database.Profile
-import com.github.shadowsocks.database.ProfileManager
-import com.github.shadowsocks.plugin.AlertDialogFragment
-import com.github.shadowsocks.plugin.Empty
 import com.github.shadowsocks.preference.DataStore
 import com.github.shadowsocks.preference.OnPreferenceDataStoreChangeListener
 import com.github.shadowsocks.utils.Key
@@ -63,7 +54,6 @@ import com.github.shadowsocks.widget.ServiceButton
 import com.github.shadowsocks.widget.StatsBar
 import com.google.android.material.navigation.NavigationView
 import com.google.android.material.snackbar.Snackbar
-import kotlinx.android.parcel.Parcelize
 
 class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPreferenceDataStoreChangeListener,
         NavigationView.OnNavigationItemSelectedListener {
@@ -72,17 +62,6 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
         private const val REQUEST_CONNECT = 1
 
         var stateListener: ((BaseService.State) -> Unit)? = null
-    }
-
-    @Parcelize
-    data class ProfilesArg(val profiles: List<Profile>) : Parcelable
-    class ImportProfilesDialogFragment : AlertDialogFragment<ProfilesArg, Empty>() {
-        override fun AlertDialog.Builder.prepare(listener: DialogInterface.OnClickListener) {
-            setTitle(R.string.add_profile_dialog)
-            setPositiveButton(R.string.yes) { _, _ -> arg.profiles.forEach { ProfileManager.createProfile(it) } }
-            setNegativeButton(R.string.no, null)
-            setMessage(arg.profiles.joinToString("\n"))
-        }
     }
 
     // UI
@@ -186,29 +165,6 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
         changeState(BaseService.State.Idle) // reset everything to init state
         connection.connect(this, this)
         DataStore.publicStore.registerChangeListener(this)
-
-        val intent = this.intent
-        if (intent != null) handleShareIntent(intent)
-    }
-
-    override fun onNewIntent(intent: Intent) {
-        super.onNewIntent(intent)
-        handleShareIntent(intent)
-    }
-    private fun handleShareIntent(intent: Intent) {
-        val sharedStr = when (intent.action) {
-            Intent.ACTION_VIEW -> intent.data?.toString()
-            NfcAdapter.ACTION_NDEF_DISCOVERED -> {
-                val rawMsgs = intent.getParcelableArrayExtra(NfcAdapter.EXTRA_NDEF_MESSAGES)
-                if (rawMsgs != null && rawMsgs.isNotEmpty()) String((rawMsgs[0] as NdefMessage).records[0].payload)
-                else null
-            }
-            else -> null
-        }
-        if (sharedStr.isNullOrEmpty()) return
-        val profiles = Profile.findAllUrls(sharedStr, Core.currentProfile?.first).toList()
-        if (profiles.isEmpty()) snackbar().setText(R.string.profile_invalid_input).show()
-        else ImportProfilesDialogFragment().withArg(ProfilesArg(profiles)).show(supportFragmentManager, null)
     }
 
     override fun onPreferenceDataStoreChanged(store: PreferenceDataStore, key: String) {
